cgame(C6b): two sign-of-zero classes — encoded-bbox int negation + oracle cexp fusion
Encoded bbox (702 findings): Raven declares x/zd/zu as int in
CG_ClipMoveToEntities (cg_predict.c:218), so -x of a zero solid byte is
integer 0 and reaches CM_TempBoxModel as +0.0. The port negated floats
and sent -0.0. The port now negates ints, matching Raven.

Shield axis (509 findings): the module sent axis[0][1] = -0.0 where the
recording holds +0.0. The chain (bytedirs y +0.0 -> VectorScale -1 ->
vectoangles yaw -0.0 -> AngleVectors sy) is -0.0 under IEEE in both C
and Rust - the recording's +0.0 is an oracle BUILD artifact: GCC -O2
fuses the same-argument sin+cos pair in AngleVectors into one
__builtin_cexpi call, that lands on Apple's cexp, and Apple's cexp
returns +0.0 imag for a -0.0 angle (repro: linked q_math.o showed
_cexp relocations and +0.0; step-by-step sin gives -0.0). The Rust port
was already faithful. build.sh now passes -fno-builtin-sin/-cos (and
the f/sincos variants), which restores real libm calls and -0.0. The
swoop1 trace must be re-recorded with the rebuilt dylib before this
class disappears from the referee.

Both temporary RSDBG probes are removed.

+++ crates/mp/cgame/src/cg_predict.rs
@@ -689,12 +689,14 @@ pub fn CG_ClipMoveToEntities(
             origin = o;
         } else {
             // encoded bbox
-            let x = (solid & 255) as f32;
-            let zd = ((solid >> 8) & 255) as f32;
-            let zu = (((solid >> 16) & 255) - 32) as f32;
+            // Raven negates x/zd as ints, so a zero stays +0.0 - a float
+            // negation gives -0.0 and the C6b referee sees the sign bit.
+            let x = solid & 255;
+            let zd = (solid >> 8) & 255;
+            let zu = ((solid >> 16) & 255) - 32;
 
-            let mut bmins: vec3_t = [-x, -x, -zd];
-            let mut bmaxs: vec3_t = [x, x, zu];
+            let mut bmins: vec3_t = [(-x) as f32, (-x) as f32, (-zd) as f32];
+            let mut bmaxs: vec3_t = [x as f32, x as f32, zu as f32];
 
             let eType = ctx.world.entities[num].currentState.eType;
             let npcClass = ctx.world.entities[num].currentState.NPC_class;
